services/shared/requirements.txt
# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4
redis==5.0.1

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0

# ML/AI для RAG (ОБЯЗАТЕЛЬНО для функционала!)
sentence-transformers==4.1.0
transformers>=4.41.0
huggingface_hub>=0.19.4
numpy>=1.24.3

# HTTP clients
httpx==0.25.2
requests==2.31.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6 